move client activity to ensure public api is legit
